// src/services/models/pageantdetails/pageantTitleData.ts
import {Base} from '../base';

export interface PageantTitleData extends Base<Data> {}

export interface Data {
  pageants: Pageants;
}

export interface Pageants {
  current_page: number;
  data: Title[];
  first_page_url: string;
  from: number;
  last_page: number;
  last_page_url: string;
  next_page_url: string;
  path: string;
  per_page: number;
  prev_page_url: any;
  to: number;
  total: number;
}

export interface Title {
  title: string;
  id: number;
}